src/converter-service/requirements.txt
pymongo==4.15.3
fastapi==0.120.2
python-dotenv==1.2.1
uvicorn==0.38.0
pika==1.3.2